Added minor changes to cashierLogin
Renamed file to cashierLogin
Added extra j label
