+++ nbproject/private/private.xml
@@ -4,11 +4,7 @@
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
             <file>file:/C:/Users/kithu/Documents/GitHub/SupraMart-System/src/lk/supramart/gui/admin/adminDashboard.java</file>
-
-            <file>file:/D:/College/Projects/SupraMart-System/src/lk/supramart/cashier/login.java</file>
-            <file>file:/D:/College/Projects/SupraMart-System/src/lk/supramart/gui/Home.java</file>
-            <file>file:/D:/College/Projects/SupraMart-System/src/lk/supramart/gui/SplashScreen.java</file>
-            <file>file:/D:/College/Projects/SupraMart-System/src/lk/supramart/gui/admin/AddSupplier.java</file>
+            <file>file:/C:/Users/kithu/Documents/GitHub/SupraMart-System/src/lk/supramart/gui/admin/addProduct.java</file>
         </group>
     </open-files>
 </project-private>
